test: refactor cart items request specs
